Move JsonMessageQuery to rest-api-resources module
